Alter sbw_einzugsgebiet Filter (#53)
the changes in the sbw_einzugsgebiet view altered the geometry type to MultiSurfaces. When using ST_Dump on a ST_Multisurface that has only one part, the dump cascades through to the Exterior Ring, resulting in a CoumpoundCurve (in PostGIS 3.0.0). This PR alters the perimeter query to pass CurvePolygons regardless of the structure