tww-project 20250912
The actual version of the project file, based on database2025.0.2rc6
I think the layers are 98-99% configured like this (also the descriptions).

Im Teil Rural bin ich nicht überall sicher, wie das genau zusammenpassen soll mit farm, building_group und disposal.

I think the project is ready to translate. There will probably be some minor correction in the rural-group, but this will not concern the translations.

Would be great, if we can work soon with tww2025 and if we can translate the project-file soon (but I have holidays in week 38 and 39).